refactor: move credentials management UI into a modal
Extract OAuth/static credential rows from inline rendering in
ConnectorCredentialsSection into a ManageCredentialsDialog, triggered
by a "Manage credentials" button. Keeps the Agent Credential dropdown
inline while reducing visual clutter on the connector config page.
